+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/CrossbowItemMixin.java
@@ -1,7 +1,7 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
-import insane96mcp.iguanatweaksreborn.module.combat.stats.Stats;
+import insane96mcp.iguanatweaksreborn.module.combat.MiscStats;
 import net.minecraft.world.item.CrossbowItem;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -10,6 +10,6 @@ import org.spongepowered.asm.mixin.injection.At;
 public class CrossbowItemMixin {
     @ModifyExpressionValue(method = "getShootingPower", at = @At(value = "CONSTANT", args = "floatValue=3.15"))
     private static float powerBonusFlat(float original) {
-        return Stats.crossbowVelocity.floatValue();
+        return MiscStats.crossbowVelocity.floatValue();
     }
 }