+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/DiggerItemMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
-import insane96mcp.iguanatweaksreborn.module.combat.stats.Stats;
+import insane96mcp.iguanatweaksreborn.module.combat.MiscStats;
 import insane96mcp.iguanatweaksreborn.module.mining.Gold;
 import insane96mcp.insanelib.base.Feature;
 import net.minecraft.world.item.DiggerItem;
@@ -21,6 +21,6 @@ public class DiggerItemMixin {
 
 	@ModifyConstant(method = "hurtEnemy", constant = @Constant(intValue = 2, ordinal = 0))
 	public int onHurtEnemy(int hurtAmount) {
-		return Feature.isEnabled(Stats.class) && Stats.oneDamageForToolAttacking ? 1 : hurtAmount;
+		return Feature.isEnabled(MiscStats.class) && MiscStats.oneDamageForToolAttacking ? 1 : hurtAmount;
 	}
 }