+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/MobEffectMixin.java
@@ -1,7 +1,7 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
-import insane96mcp.iguanatweaksreborn.module.combat.stats.Stats;
+import insane96mcp.iguanatweaksreborn.module.combat.MiscStats;
 import insane96mcp.iguanatweaksreborn.module.misc.tweaks.Tweaks;
 import insane96mcp.insanelib.base.Feature;
 import net.minecraft.world.effect.MobEffect;
@@ -18,14 +18,14 @@ public class MobEffectMixin {
 
 	@ModifyExpressionValue(method = "applyEffectTick", at = @At(value = "CONSTANT", args = "intValue=4"))
 	public int onHealFactorTick(int healFactor) {
-		if (Feature.isEnabled(Stats.class) && Stats.betterHealingPotion)
+		if (Feature.isEnabled(MiscStats.class) && MiscStats.betterHealingPotion)
 			return 6;
 		return healFactor;
 	}
 
 	@ModifyExpressionValue(method = "applyInstantenousEffect", at = @At(value = "CONSTANT", args = "intValue=4"))
 	public int onHealFactorInstantaneous(int healFactor) {
-		if (Feature.isEnabled(Stats.class) && Stats.betterHealingPotion)
+		if (Feature.isEnabled(MiscStats.class) && MiscStats.betterHealingPotion)
 			return 6;
 		return healFactor;
 	}