+++ src/main/java/insane96mcp/iguanatweaksreborn/module/hungerhealth/healthregen/HealthRegen.java
@@ -113,7 +113,7 @@ public class HealthRegen extends Feature {
 	}
 
 	/**
-	 * Different from Player#isHurt as doesn't return true if missing less than half a heart
+	 * Different from PlayerStats#isHurt as doesn't return true if missing less than half a heart
 	 */
 	public static boolean isPlayerHurt(Player player) {
 		return player.getHealth() > 0 && player.getHealth() <= player.getMaxHealth() - 1;