+++ changelog.md
@@ -10,6 +10,8 @@
   * Client
     * World border has its own feature now
     * Death has its own feature now
+  * Combat
+    * Player stats has its own feature now
   * Cleaned up some config descriptions
 
 ## 3.21.5